+++ TrelloDotNet/TrelloDotNet/AutomationEngine/Model/Triggers/LabelRemovedFromCardTrigger.cs
@@ -7,12 +7,12 @@ using TrelloDotNet.Model.Webhook;
 namespace TrelloDotNet.AutomationEngine.Model.Triggers
 {
     /// <summary>
-    /// Trigger that occurs when a Card has a label removed
+    /// A Trigger that occurs when a Card has a label removed
     /// </summary>
     public class LabelRemovedFromCardTrigger : IAutomationTrigger
     {
         /// <summary>
-        /// Set this to 'True' if you supplied names of labels instead of the Ids. While this is more convenient, it will in certain cases be slightly slower and are less resilient to renaming of things.
+        /// Set this to 'True' if you supplied the names of labels instead of the Ids. While this is more convenient, it will in certain cases be slightly slower and less resilient to the renaming of things.
         /// </summary>
         public bool TreatLabelNameAsId { get; set; }